requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.6.2.post1
blis==1.0.1
catalogue==2.0.10
click==8.1.7
cloudpathlib==0.20.0
confection==0.1.5
configobj==5.0.9
configparser==7.1.0
contourpy==1.3.0
cycler==0.12.1
cymem==2.0.8
dnspython==2.7.0
etelemetry==0.3.1
exceptiongroup==1.2.2
fitz==0.0.1.dev2
fonttools==4.54.1
frontend==0.0.3
h11==0.14.0
httplib2==0.22.0
importlib_resources==6.4.5
isodate==0.6.1
itsdangerous==2.2.0
kiwisolver==1.4.7
langcodes==3.4.1
language_data==1.2.0
looseversion==1.3.0
lxml==5.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
matplotlib==3.9.2
mdurl==0.1.2
mkl-service==2.4.0
murmurhash==1.0.10
nibabel==5.3.1
nipype==1.8.6
packaging==24.1
pandas==2.2.3
pathlib==1.0.1
preshed==3.0.9
prov==2.0.1
psutil==6.1.0
pydantic==2.9.2
pydantic_core==2.23.4
pydot==3.0.2
Pygments==2.18.0
pymongo==4.10.1
pyparsing==3.2.0
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pyxnat==1.6.2
rdflib==6.3.2
rich==13.9.2
scipy==1.14.1
shellingham==1.5.4
simplejson==3.19.3
six==1.16.0
smart-open==7.0.5
sniffio==1.3.1
spacy==3.8.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
starlette==0.41.0
thinc==8.3.2
tqdm==4.66.5
traits==6.3.2
triton==3.1.0
typer==0.12.5
tzdata==2024.2
uvicorn==0.32.0
wasabi==1.1.3
weasel==0.4.1
wrapt==1.16.0
Brotli==1.0.9
certifi==2023.7.22
charset-normalizer==3.3.0
filelock==3.13.1
# gmpy2==2.1.05
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.3
mpmath==1.3.0
numpy==2.0.1
pillow==10.1.0
PySocks==1.7.1
PyYAML==6.0
requests==2.31.0
sympy==1.13.1
typing_extensions==4.8.0
urllib3==2.1.0

